requirements.txt
pdfplumber
pymongo
python-dotenv
pandas
ollama
pikepdf
requests
msal
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
py-mon
pymongo-amplidata